refactor(cloud): extract OpenAI wire parsing helpers into dedicated namespace
Move the OpenAI Chat Completions field-level parsers (parseToken,
parseReasoningDelta, parseToolCallDeltas, parseWholeToolCalls,
parseFinishReason, parseUsage, parsePrefillProgress) and their nested
support types (ToolCallDelta, WholeToolCall, PrefillProgress) out of
OpenAIBackend.swift into a dedicated OpenAIChatCompletionsPayloadParsing
namespace file.

OpenAIBackend now sheds the wire-format vocabulary it shouldn't own — its
responsibility is request building and lifecycle. The parsers live next
to OpenAIStreamEventExtractor, which is their only stateful consumer.
Also collapses the dispatch shim in CloudPayloadHandler.swift that
previously bounced through legacyExtract{Token,Usage} on the backend.

OpenAIBackend.swift: 631 -> 426 LOC.

Architect review of PR #1272 (Phase 2/B/iii/delta) flagged this as
follow-up A.
